+++ src/appleseed/renderer/kernel/lighting/pt/ptlightingengine.cpp
@@ -453,7 +453,7 @@ namespace
         const LightSampler&     m_light_sampler;
 
         uint64                  m_path_count;
-        Population<size_t>      m_path_length;
+        Population<uint64>      m_path_length;
     };
 }
